Rename WasdKeyboardAdapter.java
